watch: split webhook logic out of check.go, Store out of watch.go
Resolves the merge conflict between main's SSRF-hardened webhook client
(PR #508) and release/1.21.0's version of check.go/watch.go, and brings
both files under the 800-line ceiling:

- internal/watch/webhook.go (new): fireWebhook, the SSRF-safe HTTP
  client, and webhook payload/redaction helpers, extracted from
  check.go (437 lines, was ~550+).
- internal/watch/store.go (new): the Store type and all persistence/
  route-pruning methods, extracted from watch.go (293 lines, was 835).
- webhook_log_test.go: fixed TestFireWebhookCreateRequestFailureDoesNotLogTheSecret,
  which targeted the pre-hardening code path. fireWebhook's up-front
  scheme/parse guard now catches a DEL-byte URL before
  http.NewRequestWithContext is ever reached, so the test now asserts
  on the guard's rejection line instead (still verifies no secret leak).
- trvl_home_disclosure_test.go: repointed the internal/watch/watch.go
  ~/.trvl-writer registry entry to store.go, where the writer now lives.

go build/vet/test clean on everything touched. 5 unrelated pre-existing
failures (dealquality x3, TestGridCmd x2) verified present at clean HEAD
via git stash -u, untouched by this change.
